+++ src/kappaml_core/meta/__init__.py
@@ -2,8 +2,10 @@
 The :mod:`kappaml_core.meta` module contains meta-learning algorithms
 """
 
+from .meta_classifier import MetaClassifier
 from .meta_regressor import MetaRegressor
 
 __all__ = [
     "MetaRegressor",
+    "MetaClassifier",
 ]